Substitution example: removed hedging, added links
I revised the text after realizing that there actually are keys `:field`
and (in Vega-Lite) `"field"`.  I thought `:field` might just be a dummy
variable for the sake of an example.

I also added links to Hanami source and a Vega-Lite doc page.  Maybe
it's inappropriate to include these links in a section that's intended
to give an overview.  However, part of what's confusing is all of the
layers involved (:X is value of :field, but it's a key to look up "x",
which will be a default value for "field").  One doesn't have to follow
the links, but maybe having the exact references available will help
some readers to sort out the layers.

Fixed typos created when fixing typos